backend/requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
ecdsa==0.19.1
eventlet==0.40.3
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyasn1==0.6.1
python-dotenv==1.2.1
python-engineio==4.12.3
python-jose==3.5.0
python-socketio==5.14.3
redis==7.1.0
requests==2.32.5
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.3.2
